# openapi.json
{"openapi":"3.0.1","info":{"title":"User and Github","description":"It provides operations of Repository and User. And it provides APIs for manipulating CI/CD pipelines.","version":"0.1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/works/{workId}":{"get":{"tags":["work-controller"],"description":"Get information of a work.","operationId":"getWork","parameters":[{"name":"workId","in":"path","description":"A work id","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"A requested work object was not found.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Work"}}}},"200":{"description":"A requested work object","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Work"}}}}}},"put":{"tags":["work-controller"],"description":"Stop a specific work by passing status property.","operationId":"updateWork","parameters":[{"name":"workId","in":"path","description":"A work id","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateWork"}}},"required":true},"responses":{"200":{"description":"A status has been updated successfully.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Work"}}}}}}},"/user/{userId}/repositories/{repositoryId}":{"put":{"tags":["user-controller"],"description":"Enable or disable CI/CD on this repository.","operationId":"updateRepository","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string"}},{"name":"repositoryId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRepository"}}},"required":true},"responses":{"200":{"description":"This repository has been updated successfully.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Repository"}}}}}}},"/works":{"get":{"tags":["work-controller"],"description":"Get a list of previous works following search conditions.","operationId":"getWorkList","parameters":[{"name":"from","in":"query","description":"from time","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"to","in":"query","description":"to time","required":true,"schema":{"type":"string","format":"date-time"}}],"responses":{"200":{"description":"A list of works","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Work"}}}}}}},"post":{"tags":["work-controller"],"description":"Start new CI/CD work for a specific commit.","operationId":"createWork","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWork"}}},"required":true},"responses":{"200":{"description":"New work is created and started successfully.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Work"}}}},"500":{"description":"A work creation failed.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Work"}}}}}}},"/user/{userId}":{"get":{"tags":["user-controller"],"description":"Get user information of a user id.","operationId":"getUser","parameters":[{"name":"userId","in":"path","description":"An id where you want to get information","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"A status has been updated successfully.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}},"404":{"description":"Requested user information was not found.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/user/{userId}/repositories":{"get":{"tags":["user-controller"],"description":"Get a list of user's repositories.","operationId":"getRepositories","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"A requested list","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Repository"}}}}}}}}},"components":{"schemas":{"UpdateWork":{"type":"object","properties":{"status":{"type":"string","enum":["STARTED","STOPPED","FAILED","COMPLETED"]}},"description":"Object containing status you expect as new status"},"Work":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"status":{"type":"string","enum":["STARTED","STOPPED","FAILED","COMPLETED"]},"url":{"type":"string"},"repository_id":{"type":"string"},"commit_id":{"type":"string"},"commit_message":{"type":"string"},"started_time":{"type":"string","format":"date-time"},"completed_time":{"type":"string","format":"date-time"},"failed_time":{"type":"string","format":"date-time"},"run_script":{"type":"string"}}},"UpdateRepository":{"type":"object","properties":{"enabled":{"type":"boolean"}},"description":"object containing whether CI/CD is enabled"},"Repository":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"visibility":{"type":"string"},"id":{"type":"string"},"enabled":{"type":"boolean"},"url":{"type":"string"},"html_url":{"type":"string"},"active_work_list":{"type":"array","items":{"$ref":"#/components/schemas/Work"}}}},"CreateWork":{"type":"object","properties":{"repository_id":{"type":"string"},"commit_id":{"type":"string"}},"description":"Object containing a specific commit"},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"company":{"type":"string"},"email":{"type":"string"},"avatar_url":{"type":"string"},"html_url":{"type":"string"}}}}}}